+++ db/db.go
@@ -502,10 +502,10 @@ func (db database) GetListedPosts(r *http.Request) ([]PeopleExtra, error) {
 	return ms, result.Error
 }
 
-func (db database) GetAssignedBounties(pubkey string) ([]Bounty, error) {
-	ms := []Bounty{}
+func (db database) GetAssignedBounties(pubkey string) ([]BountyData, error) {
+	ms := []BountyData{}
 
-	err := db.db.Raw(`SELECT * FROM bounty WHERE assignee = '` + pubkey + `' ORDER BY id DESC`).Find(&ms).Error
+	err := db.db.Raw(`SELECT body.*, body.id as bounty_id, body.description as bounty_description, body.created as bounty_created, body.updated as bounty_updated,  person.*, person.owner_alias as assignee_alias, person.id as assignee_id, person.description as assignee_description, person.created as assignee_created, person.updated as assignee_updated, owner.id as bounty_owner_id, owner.uuid as owner_uuid, owner.owner_pub_key as owner_key, owner.owner_alias as owner_alias, owner.description as owner_description, owner.price_to_meet as owner_price_to_meet, owner.unique_name as owner_unique_name, owner.tags as owner_tags, owner.img as owner_img, owner.created as owner_created, owner.updated as owner_updated, owner.last_login as owner_last_login, owner.owner_route_hint as owner_route_hint, owner.owner_contact_key as owner_contact_key FROM public.bounty AS body LEFT OUTER JOIN public.people AS person ON body.assignee = person.owner_pub_key LEFT OUTER JOIN public.people as owner ON body.owner_id = owner.owner_pub_key WHERE body.assignee = '` + pubkey + `' ORDER BY body.id DESC`).Find(&ms).Error
 
 	return ms, err
 }
