+++ frontend/app/src/config/host.ts
@@ -3,7 +3,8 @@ const externalDockerHosts = ['localhost:23007', 'localhost:23000'];
 
 export function getHost(): string {
   const host = window.location.host.includes('localhost') ? 'localhost:5002' : window.location.host;
-  return host;
+  return "localhost:5005";
+  // return host;
 }
 
 export function getHostIncludingDockerHosts() {
