+++ frontend/app/src/store/main.ts
@@ -729,11 +729,29 @@ export class MainStore {
       sortBy: 'created'
     });
     try {
-      const ps1 = await api.get(query);
+      const ps2 = await api.get(query);
+      const ps3: any[] = [];
 
-      this.setPersonWanteds(ps1);
+      if(ps2 && ps2.length) {
+        for (let i = 0; i < ps2.length; i++) {
+          const bounty = { ...ps2[i].bounty };
+          let assignee;
+          const owner = { ...ps2[i].owner };
 
-      return ps1;
+          if (bounty.assignee) {
+            assignee = { ...ps2[i].assignee };
+          }
+
+          ps3.push({
+            body: { ...bounty, assignee: assignee || '' },
+            person: { ...owner, wanteds: [] } || { wanteds: [] }
+          });
+        }
+      }
+
+      this.setPersonWanteds(ps3);
+
+      return ps3;
     } catch (e) {
       console.log('fetch failed getPeopleWanteds: ', e);
       return [];
@@ -751,19 +769,21 @@ export class MainStore {
       const ps2 = await api.get(query);
       const ps3: any[] = [];
 
-      for (let i = 0; i < ps2.length; i++) {
-        const bounty = { ...ps2[i].bounty };
-        let assignee;
-        const owner = { ...ps2[i].owner };
+      if(ps2 && ps2.length) {
+        for (let i = 0; i < ps2.length; i++) {
+          const bounty = { ...ps2[i].bounty };
+          let assignee;
+          const owner = { ...ps2[i].owner };
 
-        if (bounty.assignee) {
-          assignee = { ...ps2[i].assignee };
+          if (bounty.assignee) {
+            assignee = { ...ps2[i].assignee };
+          }
+
+          ps3.push({
+            body: { ...bounty, assignee: assignee || '' },
+            person: { ...owner, wanteds: [] } || { wanteds: [] }
+          });
         }
-
-        ps3.push({
-          body: { ...bounty, assignee: assignee || '' },
-          person: { ...owner, wanteds: [] } || { wanteds: [] }
-        });
       }
 
       return ps3;
